# pkg/api/types.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen/v2 version v2.0.0 DO NOT EDIT.
package api

// Pagination defines model for Pagination.
type Pagination struct {
	Limit  *int `json:"limit,omitempty"`
	Offset *int `json:"offset,omitempty"`
	Total  *int `json:"total,omitempty"`
}

// Workload defines model for Workload.
type Workload struct {
	Description  *string `json:"description,omitempty"`
	Envinronment *string `json:"envinronment,omitempty"`
	Id           *string `json:"id,omitempty"`
	Name         *string `json:"name,omitempty"`
}

// AddWorkloadJSONRequestBody defines body for AddWorkload for application/json ContentType.
type AddWorkloadJSONRequestBody = Workload